+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2025-03-12
+
+### Changed
+- Simplified post-install message for improved installation speed
+- Added timeout to post-install script to prevent installation delays
+
 ## [0.2.3] - 2025-03-12
 
 ### Fixed